Product class added
