+++ .gitignore
@@ -2,3 +2,4 @@
 /.settings/
 /.classpath
 /.project
+/.idea/workspace.xml
