+++ .idea/libraries/Maven__org_easytesting_fest_assert_core_2_0M10.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.easytesting:fest-assert-core:2.0M10">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/easytesting/fest-assert-core/2.0M10/fest-assert-core-2.0M10.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/easytesting/fest-assert-core/2.0M10/fest-assert-core-2.0M10-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/easytesting/fest-assert-core/2.0M10/fest-assert-core-2.0M10-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
