+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Offer.java
@@ -4,11 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Offer {
-private List<OfferItem> availabeItems = new ArrayList<OfferItem>();
-	
+
+	private List<OfferItem> availabeItems = new ArrayList<OfferItem>();
 	private List<OfferItem> unavailableItems = new ArrayList<OfferItem>();
-	
-	
+
 	public Offer(List<OfferItem> availabeItems, List<OfferItem> unavailableItems) {
 		this.availabeItems = availabeItems;
 		this.unavailableItems = unavailableItems;
@@ -59,7 +58,7 @@ private List<OfferItem> availabeItems = new ArrayList<OfferItem>();
 			return false;
 		
 		for (OfferItem item : availabeItems) {
-			OfferItem sameItem = seenOffer.findItem(item.getProductId());
+			OfferItem sameItem = seenOffer.findItem(item.getProduct().getProductId());
 			if (sameItem == null)
 				return false;
 			if (!sameItem.sameAs(item, delta))
@@ -71,11 +70,10 @@ private List<OfferItem> availabeItems = new ArrayList<OfferItem>();
 
 	private OfferItem findItem(String productId) {
 		for (OfferItem item : availabeItems){
-			if (item.getProductId().equals(productId))
+			if (item.getProduct().getProductId().equals(productId))
 				return item;
 		}
 		return null;
 	}
-	
 
 }
